Use assertions to document important assumptions in the code
The GUI depends on there being a running Duke instance in order to run
commands in Duke and get proper responses.

We want to make sure that the program has such a running instance other-
wise there may be unexpected or dangerous behaviour if the GUI is run.

Use Assert statements to ensure that the GUI has a functional Duke
instance for the entire program to run correctly.

Assert statements can allow the program to crash if there happens to
not be a running Duke instance which is safer than allowing the program
to continue running.