Use Assertions
Classes do not have any checks for their invariants or
assumptions made.

Not checking for these invariants or assumptions will
cause the program to crash or give incorrect results

Add assertions to check to invariants.

Assertions is preferred as it gets the job done and it
can be turned on or off.