Refactor code and improve code quality
Several classes and methods had very long methods and made use to many
complicated expressions in the code.

This made the code less readable and more difficult to understand.

Let's refactor the code by,
* Using less complicated expressions
* Shortening methods
* Getting rid of deep nesting and arrowhead style code
* Extracting magic numbers into constants
* Making the code more explicit
* Keeping the code simple
* Improving comments for the reader

This can improve code readability and make understanding the code much
easier for both myself in the future and later users, thereby allowing
changes to be made more easily.