Fix checkstyle issues pointed out by reviewers
Checkstyle issues present in MainWindow, AddCommand and TaskList.

This caused checkstyle to fail and led to a messy structure of code.

By adjusting the indentation and ordering of import statements for MainWindow
to the correct order, checkstyle succeded.

I missed out some comments in AddCommand and TaskList as well.

By adding the appropriate comments, the code became clearer and easier to read.