Add assertions to utility classes
Exceptions are useful for detecting user errors, but can become
clunky when used to detect programmer errors.

Add assertions to the utility classes where the bulk of the
logic resides.

This will ensure that the programmer can check if his assumptions
are consistent with the existing code and make changes where
necessary.