Add assertions
Parser and Task classes have methods that require parameters given to
be valid for successful execution.

These methods either fails or does not make sense if the given
parameters are null.

Assertions are added to them to define assumptions about the program
state to ensure that parameters are confirmed to be valid.