Add assertions to check for integer bounds and null objects.
The program currently runs on the assumption that certain objects will
never be null, or that certain indexes will be within the limit of the
number of tasks.

Assertions are assumptions that are made which, if broken, causes
the program to not run as intended. Assertions are implemented to make
sure that it is easier to debug the program in future changes.

Assertions are inserted right before points in the program where it is
critical that the assumption listed is correct. This includes the check
on:
* Whether there is no referenced task in Command.java
* Indexes to delete, mark and unmark tasks in the list

Assertions are placed directly before usage to prevent incorrect
assumptions from happening should new code modify the checked variables
afterwards. This prevents an incorrect assertion from being inserted.

More assertions to be added in the future.