Refactor out Task.fromFlatpack
Storage::loadList handles part of the correct logic for loading a
packed task list, which violates SLAP (GLAP?) and causes the function
to exceed the recommended length.

Let's refactor out the code that generates the Task of the correct type
into a static method of Task where it belongs.

Readers: the term "flatpack" refers to a losslessly restorable
serialization of a data structure into a flat list, as well as the
action of performing such a serialization.