Duke.java: Refactor program flow to remove explict call to readOutput
Everytime Duke's feedback is required, an explicit call to readOutput
method is required to read the output buffer after running a command.
This design flow causes code duplication as after executing a command,
readOutput will always be called anyways.

Let's modify the program flow such that initialization of Duke and
command execution of Duke handles reading from the output stream
and instead returns the feedback string directly to the method
caller.

This program flow is preferred as to prevent accidental mishap such as
forgetting to read (and subsequently) clear the output stream after
executing a command.