Change Event date parsing to JChronic
The date parser for Event is a very basic one that can only accept ISO
8601 style strings. This is incredibly impractical for any real user.

Let's use the JChronic library to parse the date for the event instead.
This will provide recognition for all kinds of natural date inputs.

Note that this will remove support for tracking event times.
Unfortunately, I can't find a natural date and time parser that does
not use AI magic.