Add reschedule feature
The added task cannot be reschedule/modified.

Users had to delete the task then add a new task. However, this
approach will alter the sequence of the tasks in the list.

Include a reschedule command

format: reschedule {index} {Task}
This allows user to modify the saved task without changing the
sequence.