Add tests to check duplicate method
There are no unit tests to test the behaviour of new addTask method
with new function of checking duplicate tasks.

Some tests should be added to ensure that its behaviour
is consistent with what is expected. These tests can also be
used in future whenever there is a code change as a verification.

Let's add tests when duplicate todo, deadline
and event tasks are added.