Resolve checkstyle errors
GitHub prompts that some checks were not successful,
which are caused by checkstyle errors.

The errors should be fixed for successful build and check
with Gradle on GitHub as well as successful build
when checkstyle command is keyed in manually into terminal.

Let's fix checkstyle errors
to check if Gradle can build successfully now.