Implement B-Snooze for BCD-Extension
It is very inconvenient to delete a task and add a task with
same details but a different date.

Snooze command to change task to a later date.

Allow users to snooze deadline and event task to a later date.

Let's, implement a snooze command to allow users to change
their task to a later date.