Add priority level to tasks
Currently, tasks have no way to keep track of priority.

Users all have differing priority of tasks, some more important than others and it is good to capture this.

Let's implement a priority field in the tasks that will also be written into the text file when saved. This field will be of type enum.

Priority is enum so that users cannot put in unknown priority levels, for safety.