Improve code quality
Currently, there are a few functions who do not follow the principle of
single level of abstraction SLAP).

This needs to change as otherwise, it can lead to unclean code and be
messy.

Improve code quality by implementing SLAP on the code.

SLAP helps to ensure that abstraction levels are kept consistent
throughout a function and makes code easily understood.