Update Task class to include priority attribute
Task classes/objects have no priority attribute, hence it is impossible
to differentiate between tasks of different priorities

Introducing a priority attribute to tasks will allow further priority
related functions to be implemented in the future.

We introduce this attribute into the abstract Task class to make use of
inheritance, so that all inhering task-related classes will have access
to the priority attribute.