Parser class: added a try-catch block to handle the IndexOutOfBounds excpetion in the parseDelete function when the user attempts to delete a task with an index that does not exist in the list.
The user will not be able to know why he/she is unable to
input the command when this exception is not handled.

A try-catch block is added to catch the exception when it is
encountered and a message telling the user what is
the problem for him/she to diagnose the issue.

A try-catch block is the easiest and simplest way to handle such
exceptions