Add automated testing using GitHub Actions
Testing the application's output needs to be done manually by running
the runtest scripts.

This is still a fairly tedious process and prone to forgetfulness.

Automating the process would take one task off our hands.

Let's use GitHub Actions to automatically test the application when
pushing or making a pull request to the repo.

This has the added benefit of allowing testing across different
platforms.