GUI feature: make dialog boxes be able to show the full list for long lists
GUI only showed truncated input when the task list was long.

Modified the Label height in order to accommodate longer text.

Modifying JavaFX label properties is preferable to getting the user to
maximise the application window or to leave the message truncated.

Utilised this Stack Overflow discussion to make modifications:
https://stackoverflow.com/questions/35114338/how-do-i-make-my-vbox-
and-label-resize-according-to-texts-in-label-in-javafx