Add assertions
There are several assumptions made in the code that are not verified.

The assumptions may result in unexpected errors or incorrect output
if they are not confirmed to be accurate.

Add assertions at appropriate points to confirm if assumptions made
are correct, otherwise, the assertion would fail.

Utilising assertions would allow the programmer to realise a mistake
was made in the code.