Use assert feature
The assert feature is used to define assumptions about the program state so that the runtime can verify them.

Implement the assert features in methods where the execution of the method have to be of certain assumption before running.

This is to allow the program to run as per design and to catch any possible bugs and terminate the program as soon as it detects.