Improve code quality
The default file path had to be explicitly stated when creating the Duke object and this requires the user to know the file path.

Store the default file path as a named constant.

By storing it as a named constant, the user does not need to know the file path when initialising the object. The Duke class will by default create the file in the default file path if no file path is specified.