Add update feature to tasks
Users are not able to rename the task that they added to the list.

Add a new update feature to allow users to change the name of the task.

This allow the users to easily update the task instead of having to delete the task and add a new task.