refactor: add assertions
Classes in the command package accept any string as the command when
they are initialised.

This could result in a wrong command being created if the wrong
constructor is called for a certain command string (e.g. a DeleteCommand
could possibly be created for "todo abc", which should be an
AddTodoCommand).

Assertions are added to all factory methods and constructors that assert
what their command strings should start with or be equal to.

Since Parser should already ensure that the correct constructors and
factory methods are called for a certain command string, asserts ensure
that this does not change in future updates in code.