Fix launcher
Previously, the poolsheen.jar file could not be executed because
the mainClassName was specified wrongly.

We want to fix the jar executable so that the Poolsheen program
works.

Let's change the path specified in mainClassName to
poolsheen.Launcher.

This ensures that all executables will always run from the Launcher
class first.