A-MoreErrorHandling: add more error messages
User could mark or unmark a task repeatedly without any warning
for tasks that are already marked or unmarked.

Warning messages could be displayed to improve this function
to prompt user about it.

Let's add warning messages when user wants to mark or unmark
tasks that have already been marked or unmarked previously.