Extract the findTasks component from FindCommand Class to TaskList Class
The current implementation of execute method in FindCommand iterates through
the entire TaskList, compare each task's description with the keywords before
appending it to the resulting String to return.

The iteration of the TaskList and searching for the matching Tasks should be
encapsulated within the TaskList Class.

Let's abstract out the component for searching for matching Tasks via keywords
into a new method within the TaskList Class called findTasks.

This follows after the OOP principle and prevents weak SLAP.