Update Deadline
Currently, the access modifier for date in Deadline is protected

It needs to be changed to private which is more secure than access
modifier. We do not want other classes to change this variable
directly.

Change protected access modifier to private

It is done this way to secure data within the class, and is the most
secure way to do it by the Java convention