Include Assertions in AddCommand Class and Storage Class
The else block in the execute method within AddCommand class assumes
the taskType to be 2 (which refers to Event tasks). The load method
in Storage class assumes that the directory and file are created correctly.

Validates the taskType before creating an Event task. To ensure that
the directory and file have been created before writing to or reading
from the file.

Let's,
* Include an assert statement to check for the correct taskType in the
else block of the execute method.
* Include assert statements after commands to create file and create
directory to check if the directory and file exist before proceeding.

Validate the assumptions made and terminate the program if there is
an error before proceeding on with the rest of the code.