Add assertions to Parser.java
There are several methods that require specific inputs and will fail if
these inputs are incorrect.

Using assertions will help to detect any mistakes in parsing inputs
before using them in these functions.

Let's document any assertions used in the future so that it is easy to
see what assertions have to be adhered to.