Add assert statements
Lack of assert statements might cause the developer to miss out on
certain flaws present in the program.

Several assert statements have been implemented to prevent this from
happening.

Adding assertions allows the developer to identify obvious flaws i.e
things that the programmer knows to be true about the program or state
of objects and variables.

Assertions is useful to test the code since it helps identify bugs present in the program.