Add the ability to detect duplicate tasks.
Duke chatbot is unable to detect duplicate tasks.

An automatic checker for duplicate tasks is more user-friendly
because users cannot be expected to remember what tasks they have
already added to the list.

Let's,
* Update the task adding algorithm in the TaskList class to check for
duplicate tasks
* Add an error message if the user attempts to add duplicate tasks

Modifying the TaskList class is preferable as it follows the single
responsibility principle since this class is responsible for the
handling of tasks in the list.