Replace magic number in checkStatus() method
In the checkStatus() method, integer 1 means the task is done and 0
means the task is not done yet. Simply compare the parameter with an
integer can be confusing as others don't know what's the meaning
behind the integer.

Using a named constant to represent the status of a completed task
and compare the parameter to the constant is easier to understand
because the name tells us more about the meaning of the number.

Let's declare a new constant DONE to represent the status of a completed
task and compare the parameter to DONE to determine whether the task is
completed or not.