Replace Command enum for abstract class
Enum is used for encapsulating the different commands.

The common behaviours across different commands cause code duplication.

Extracting the common behaviour into an abstract class allows for the
use of polymorphism when running a user command in the bot. For example,
when the bot is executing a command, it does not have to check what
command it is before executing. This reduces the lengthiness of the
getResponse method in the Duke class.

Let's pull up behaviours common to all command classes into a new
abstract parent class called Command.

Using inheritance is preferable over composition in this situation
because the common behaviors are not composable.