Storage load method: extract methods three helper methods createTask, createDeadline and create Event
The Storage load method involves the creation of different types of tasks,
which makes the code very long and unreadable.

Extracting the creation of each type of task allows us to abstract away the
details according to SLAP and replace the long lines of code with multiple
short methods for better readability.

Let's
* move the code fragment that creates the Task to a new method called createTask
* further split the createTask method into two methods: createEvent and createDeadline
* createEvent contains the code fragment for creating events
* createDeadline contains the code fragment for creating deadlines