Add Java assert feature
Integer variables used as counters, in some classes that inherit from
Command, cannot be less than zero.
In some classes that inherit from Command, the task description
and/or the date cannot be empty.

So let's assume that these are not the case in the code.
If the cases which they do happen, we let the runtime detect an
assertion failure to inform us that there are potentially bugs that
cause it to happen.