Add remind task functionality
Reminding tasks is a new feature implemented into this program. It
is used to tell the user which tasks need to be completed within
7 days.

This feature can be used when the user types in "remind" in the
CLI.

The goal of this command is to assist users to easily view which
tasks need to be completed within 7 days.