Add Assertions in classes Duke and Storage
Storage.decode(): Under the current format for storing tasks,
the [type] of the task must be "T", "D", or "E".
The default case for the switch should never be reached.

Duke.handleInput(): An empty input after trimming is caught
by the UI level and gracefully ignored. The input here is
guaranteed to not be an empty string.