Implement Stream for findTasks and compileTasks methods in TaskList Class
The two methods in TaskList Class iterate through the entire ArrayList(Task) to
check for keywords and add it to the compiled or matched list.

This can be done more efficiently via Streams using map and filter functions that
are under the Stream Class.

Let's convert the TaskList into a Stream<Task> and utilise the map and filter function
to perform the same task.