Reintroduce Welcome Message
Currently, the welcome message is mistakenly removed when moving to JavaFX UI.

Without it, the user starts up to an empty window, which makes Duke seem incomplete.

Reintroducing it in the MainWindow makes sense because it is the new equivalent to the old Ui class, where the message used to be displayed.

Let's
- reintroduce the welcome message in MainWindow
- add a bit of personality to it