Add sort functionality
There is no simple way to organize tasks in the task list.

Users who want to sort their tasks by upcoming deadline or alphabetical
order could not do so easily.

Let's add a sort command that allows users to sort the list either
alphabetically or chronologically, and in ascending or descending
order.

Let's also refactor the Task class to include time as well.