C-Help: add in-app guidance
There is in-app guidance of quick start or detailed guidance
when user presses the respective button, but there is no command
to the guidance if user prefers typing a command.

Other than clicking the buttons for guide,
a help command that leads to the guide could also help
users to navigate this program more easily.

Let's add a help command for users to access the detailed help page.