Add extension to detect duplicates
Currently, same tasks can be added, resulting in meaningless duplicates.

By adding this extension, users will be informed when a task already
exists and the same task will not be added more than once.

Let's detect duplicates and print a message to the user when the same
task is added.