Improve code quality
Some classes have methods that are not in used.

Those methods pile up and make it difficult for other people to read the
code.

Remove unnecessary fields and methods in classes.

Clean the code is better than comment out because the program no longer
relies on those methods and fields.