Add assertions within Parser.java and Storage.java
Parser class: add assertions to check for new command
Storage class: add assertions to check that taskArray is not empty

No checks on whether the command returned is null or the taskArray
returned is empty.

For Duke to work correctly, there should be a valid command and
populated taskArray.

Assert statements are added within the 2 classes to verify the
assumption that a valid command and populated taskArray is returned.

Assertion failure can help to highlight a mistake in the code.