Edit method length and structuring for code quality
This is done so that the overall code quality of the program is of the
expected level of the module.

Every method is limited to thirty lines and three indentation levels.
The properties of SLAP AND KISS is applied to the program.

Thes rules mentioned above are powerful indicators to check whether the
quality of the code is up to standard.

The guidelines can be found  the CS2103T textbook under code quality.