Implement C-Undo extension for the ChatBot
There are no methods to undo a command, the existing method to "undo"
requires user to manually utilise existing method to undo their previous
command.

Provide a way to undo the most recent command which allows
user to restore the ChatBot to the state before the command
was executed with a simple "undo" command.

Let's
* Create an UndoCommand Class to handle the logic of undo-ing a command
* Create a HelpCommand Class to handle the prompting of user with existing
commands and examples
* Include assert statements to test assumptions made
* Remove unnecessary methods like isExit()