Add assertions. Update the names in Storage to fit naming conventions.
Assertions were placed in the readInputString method in Parser
so that the first element in splitCommand will never be null. This
is important to check since the rest of the code depends on
what is contained in the first element of splitCommand.

Also, comments made by my peers about the naming of certain
variables and methods were heeded. I changed the name of load()
to loadFromFile() and I changed the name of File f to File taskFile.