Add exceptions
One of the most common errors that may be produced from this
application is parsing error. For example, if the user did not
enter the date and time for a Deadline and Event according to our
format.

Let's introduce a new class of Exception called MewException
to handle all possible exceptions more specifically.