Fix missing user input bug
Some commands like mark, unmark and delete require users to specify the
index of the task. Currently, if a user omit the index, there will be no
output.

A message should be sent to remind users to retype their command with an
appropriate index.

To fix the bug of a missing index, a
checkForMissingArgs() method is introduced to remind user to key in an
index for those commands should they left it out.