Improve code quality through refactoring.
The "bye" case in readInputString method was not adhering to
SLAP.

Thus, I added in two methods into taskList so that I could abstract
out what I was doing before so as to SLAP hard. This would make
it easier to understand the code.

Also, since I had not fully implemented changing of the date format,
I practiced KISSing. I made the code simple to change the date
format. I also added in new methods in the UI to show related
messages when the user want to change the date format.