Style: fix style errors and improve code quality
Code has a few style violations, such as missing punctuation at
the end of the params in the JavaDocs.

The code quality could also be improve, by enforcing the SLAP principle,
using guard clauses and improving readability.

Let's fix all the styling errors and try to ensure that code quality adheres
to the coding standard.