Add A-CodeQuality
Review code quality to adhere to coding standard

There are several parts which were not tidied up; in particular, the "//Fallthrough" comment was missing for the 'switch' statement, where I intended for it to fall through, but did not write an explicit "Fallthrough" comment. I will continue to make sure the subsequent codes closely adhere to the accorded coding standard.