Add Anomoly Detection
When a new Event is added with the exact same date as a previous Event, it will be added to the To-Do List.

It would be helpful to add an ability to check for clashes in the Event dates, so that one do not commit to multiple events on the same day.

Let's add an additional method to check for this behaviour under each Event, and throw an exception when the same date is detected.

This can allow us to detect clashes while maintaining not adding the additional Event to our list.