Add assertions to Parser methods
There are a few methods in Parser that will be invalid if the user
inputs are not in the correct format.

Using assertions to define assumptions about the program state so that the runtime can verify them allows bugs in the code to be detected.

Let's define assertions to ensure that assumptions in the program are
not violate.