Fix code quality violations
Priority when writing code is on functionality and making it work. There are several areas where the code quality is subpar.

Code quality has to be improved to increase readability and make it easier to debug hidden bugs.

Let's improve code quality by closely following the code quality standards specified.