Image retrieval: may return null.
Program may not run as intended if file is null.

Add assert to check for this.

Assert helps to check if the image is found.

Assert is done before dialogContainer is called to process the images so that it is not used unnecesarily.

Changes were made to only Duke.java and MainWindow.java.