Add assertions to verify assumptions at several places
The Duke.createTask() method may return null due to the default switch
case. This may cause errors in other parts of the code if a null value
is inserted as a Task in the TaskList.

Let's use an assertion to verify that the caller of createTask() doesn't
receive a null value by accident.

When removing a Task in the Duke.deleteTask() method, we do not check
if a null value is returned by the TaskList.removeTask() method. This
may cause an error when generating the corresponding UI messages if a
null value is used.

Let's use an assertion to verify that the returned Task object is not
null.

The String argument of the Duke.getResponse() method may be null due to
some error in the GUI classes/code. This will cause errors in other
parts of the code if a null value is used as user input.

Let's use an assertion to verify that the given String argument is not
null.