Update the welcome message in the readme
The current welcome message is a simple ASCII art
for duke.

This is outdated as the app is now called Piggy, and
the welcome message has been changes.

Let's change the code block containing the Duke ASCII
art to contain the new Piggy welcome message.

This will prevent user confusion.