Refactor the Duke.createTask() method by extracting out the code to create each type of task into separate helper methods.
The createTask() method is too long due to the code to create each type
of task. This makes the entire method hard to read and debug.

Let's extract out the code to create each type of task (Todo, Deadline
and Event) into separate helper methods. This greatly shortens the
createTask() method, making it easier to read. Furthermore, future
changes in how a task is created can now be applied to the specific
type of task instead of affecting the entire createTask() method.