Add Assertions to TaskList class
Assertions are good to detect bugs.

Let's add assertions in delete, mark and unmark methods in TaskList to
ensure that the task to be operated on is not null. Even though the methods
that are calling these methods throw exceptions when the task to be operated
is null, it is safer to use assertions as future modifications may result
in bugs that cannot be detected by the old method.