Add SortDeadline command
The deadlines is not sorted by date.

Adding a SortDeadline command can make the application more user-friendly because users now can call this command to see a list of deadlines that is sorted by dates. The deadline with the earliest deadline will be at the top of the list and the latest deadline will be at the bottom.

Let's,
* Add a TimeComparator class to compare the time.
* Add SortDeadlineCommand class, update Ui and Parser class.
* Add attribute : a list of deadlines in the taskList class and add the respective method to update and get deadlines list.