A-Assertions. Use Assertions
The code does not use assertions.

Assertions define assumptions about the program state so that the runtime can verify them.
Adding assertions can help us catch bugs by verifying our assumptions of the state
of the program.

Let's add assertions at some critical junctures of the code and verify that the runtime state
of our program coheres with our assumptions of the state of the program.

This will allow us to verify the correctness of the program, and can help us catch bugs
in the program.