Add Task Tagging Functionality
Users are unable to differentiate similar tasks with the
same attributes such as task name or deadline. Users also do not have
the functionality to add additional information to tasks outside of the
given arguments for each task type.

The lack of such functionality limit user experience which result
in potential poor user engagement and retention.

Add Tag class to represent task tags and include class field in task
classes to facilitate task tagging functionality. Modify existing
classes to account for new tag field when parsing, saving and operating
on user input strings.

Creating a Tag class allows us to abstract operations and logic on tags
in an Object-Orientated manner. This simplifies the modification to
peripheral classes and majority of the implementation logic is handled
in the Tag class.