Add assert statements.
The program have no assert statements which might cause the program to be buggy

Hence assert statements has been added following methods
parseLine() in Storage class
Add assert statements to check that the data is stored in the correct format of <type>,<mark>,<name>.

combineStringArray() in Parser class
Add assert statements to check that the start and the end parameters lie within the size of the array.

This allows for a more correct program by checking for both the preconditions.