Add Assertions for Fred.java and Storage.java
For Fred.java, add an Assertion to assert that Fred has generated a
non-empty message after receivng input from the user.

For Storage.java, add an Assertion in save method to assert that the
data file exists so that Fred can save tasks into the data file. Also
add an Assertion in load method to assert that the program only reaches
the else block if a line in the data file does not start with either
"T", "E" or "D".