Refactor code to improve code quality
1. For methods that were previously too long, additional methods
were created to replace portions of the code within the if else
blocks to allow easier debugging in the future by isolating the method
that is causing the bug.

2. Avoid complicated expression in if statements. I calculated the
intermediate values separately and include the final value for the
condition in if statement.

3. Get rid of 'magic' string that are returned from functions. Add
a variable name to the String to self explain what does the String
represent before returning it.