Add tagging to duke
Users are categorizing tasks to events, todo, and deadline.

There can be more detail added to allow the users to further categorize
their tasks.

Let's add the ability to tag tasks to duke.

Each task can have a tag to it. During construction, a default empty
tag is assigned to the task object.