C-NaturalDates. adds more flexibility to datetime parsing
bobo only accepts a very limited and rigid set of datetime formats (dd/MM/YYYY or
dd/MM/YYYY HH:mm).

The reference of dates in bobo is thus, very inflexible and not very user friendly.

Let's add a new class, DateTimeParse to handle datetime parsing. We can use
a linked hashmap to store a large variety of date or datetime formats. This hashmap
stores the regex expression corresponding to the date or datetime pattern so that
we can look through this hashmap for a matching regex expression to the datetime string,
and map it to its corresponding datetime formatter. Finally, we can parse the datetime
string into a datetime object with the formatter.

Abstracting out datetime parsing into its separate class helps enforce OOP principles
and readability of the code. Especially since it gets rather lengthy and verbose.
Abstracting it out will also make datetime parsing also more extensible so that
it is easier to add new datetime formats to be supported in the future.

However, the code could be further refactored so it is cleaner- such as using enums
and abstracting out some of the string processing into its own separate methods.