Add help command
There are many commands that can be used and each command requires the
user to follow the specified input format precisely.

New users to the app need to be able to find out what commands are available
and the specific ways to the details need to be typed in. Regular users might
also forget certain commands. Hence a help command is required.

Let's add a help command to list out the possible inputs and their required input
formats.

This allows users to quickly refer to it when needed while using Duke.