A-Assertion: Add assertion feature
Document assumptions at certain points in the code that should hold no matter what the user's input is.

Include asserts in method to ensure arguements are not null.

This will ensure subsequent methods that are linked to this are valid and do not cause further bugs

Minor javaDoc updates has been added as well