Refactor invalid task exception
The error message was being defined twice for invalid index access in both MarkCommand and UnmarkCommand

To address this, I created a new exception class to handle the error message to be generated. I also found out that the DeleteCommand was missing the index out of bounds check, so it was also added to the DeleteCommand

A new exception class was created to standardise the error message and formatting to minimise the amount of refactoring required when changing the error message content or format.