Ui: Add assertions for methods TaskList: Add assertion for getLength() method Parser: Add catch block for "mark", "unmark" and "remove" cases
For methods in Ui and TaskList, the length of the task list cannot be negative.
For parser, error is thrown when index is out of bounds.

The task list should never be negative, otherwise it indicates an error in the program.
It is more user-friendly to remind the user that the index is out of bounds rather than to terminate the entire program.

Let's
* add assertions in the various methods of Ui and getLength() in TaskList to assert that length of task list has to be positive
* add catch block for "mark", "unmark" and "remove" cases of Parser, to print an error message for the user when index is out of bounds