Update README and emoji
An emoji used to illustrate CaCa's personality could not be
displayed properly on other OS, i.e. Windows.
There is also a data folder created when users run CaCa for
the first time.

It should be changed to another emoji so that it can appear
normally as expected.
README could include more information about user's data
such as where it is saved and how user can access it.

Let's
* change the emoji to a working one
* update README.md to include more details