Add C-Sort
The user does not have the ability to sort their tasks by their deadlines.

This should be an important feature of a task-management app, in order to
enable users to quickly prioritise their to-do list.

Let's create a Sort command that prints to the user a list of their
tasks sorted in chronological order of deadlines. This list is a copy
of the original list, and does not modify the original list. All
tasks without deadlines will be placed below Deadline Tasks.

The original list should not be modified as the user may want it
sequenced in the order of addition of tasks. All tasks without deadlines
are de-prioritised as they are probably less urgent (and there
is no way to meaningfully sort tasks that have no deadlines).