Improve code quality
Some of the methods such as the parser method and the load method are
quite long and have arrowhead style code which makes it difficult to
read and understand.

Long methods have been refactored to multiple other methods. Single
level of abstraction principle implemented when appropriate.

This allow for better readibility.

There are also some changes in method names to make it less ambiguous.